refactor(compiler-cli): move defer block tests into separate file
Splits the tests for `@defer` blocks out into a separate file since the `ngtsc_spec.ts` is getting quite large.
